refactor: move support classes to the test case file
